+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+﻿## [Unreleased]
+- Diseño inicial de arquitectura
+- Configuración del repositorio y flujo Git
+- Preparación de ramas de trabajo
+
+## [v0.1.0] - Prototipo básico
+### Added
+- Estructura de carpetas base
+- Implementación de interfaces core
+- Menú de consola simple
+- Generación de organismos iniciales
+
+### Fixed
+- (Por agregar)
