+++ Core/Enums/OrganUseType.cs
@@ -1,0 +1,14 @@
+﻿namespace SimulationEvolucion.Core.Enums;
+
+public enum OrganUseType
+{
+    BuildOrgan,    // Construir órganos
+    Damage,        // Dañarse
+    Absorb,        // Absorber otros órganos
+    Heal,          // Curarse
+    Synthesize,    // Sintetizar: convertir un carácter en otro
+    Discard,       // Desechar caracteres
+    Reproduce,     // Reproducir organismo
+    Attack,        // Atacar a otros organismos
+    Defend,
+}
