+++ Core/Interfaces/IOrganUse.cs
@@ -1,0 +1,6 @@
+﻿namespace SimulationEvolucion.Core.Interfaces;
+
+public interface IOrganUse
+{
+    
+}
