+++ Core/Interfaces/ISimulationEnvironment.cs
@@ -1,0 +1,6 @@
+﻿namespace SimulationEvolucion.Core.Interfaces;
+
+public interface ISimulationEnvironment
+{
+    
+}
