+++ Core/Models/OrganUse.cs
@@ -1,17 +1,6 @@
-﻿namespace SimulationEvolucion.Models;
+﻿using SimulationEvolucion.Core.Enums;
 
-public enum OrganUseType
-{
-    BuildOrgan,    // Construir órganos
-    Damage,        // Dañarse
-    Absorb,        // Absorber otros órganos
-    Heal,          // Curarse
-    Synthesize,    // Sintetizar: convertir un carácter en otro
-    Discard,       // Desechar caracteres
-    Reproduce,     // Reproducir organismo
-    Attack,        // Atacar a otros organismos
-    Defend,
-}
+namespace SimulationEvolucion.Core.Models;
 
 public class OrganUse
 {
